Use the auto-tune version. And various fixes adopting to JtR "standard".
