+++ src/opencl_bitlocker_fmt_plug.c
@@ -24,16 +24,17 @@ john_register_one(&fmt_opencl_bitlocker);
 #else
 
 #include <string.h>
+#include <time.h>
+#include <ctype.h>
+
 #include "common-opencl.h"
 #include "opencl_device_info.h"
 #include "config.h"
 #include "options.h"
 #include "misc.h"
-#include <time.h>
-#include <ctype.h>
 
 #define CIPHERTEXT_LENGTH   64
-#define BINARY_SIZE     32
+#define BINARY_SIZE     0
 #define BENCHMARK_COMMENT   ""
 #define BENCHMARK_LENGTH    -1
 #define FORMAT_LABEL            "bitlocker-opencl"
@@ -44,10 +45,13 @@ john_register_one(&fmt_opencl_bitlocker);
 
 #define BITLOCKER_JTR_HASH_SIZE 45
 #define BITLOCKER_JTR_HASH_SIZE_CHAR 77
-
+/*
 #define MAX_PASSWORD_THREAD 8
 #define MIN_KEYS_PER_CRYPT 1
 #define MAX_KEYS_PER_CRYPT 172032
+*/
+#define MIN_KEYS_PER_CRYPT	1 /* These will change in init() */
+#define MAX_KEYS_PER_CRYPT	1
 /*
  * On a GeForce Titan X: Assuming 896 threads for 24 SMs,
  * 8 password for each thread -> 896x24x8
@@ -83,55 +87,195 @@ john_register_one(&fmt_opencl_bitlocker);
 #define UINT32_C(c) c ## UL
 #endif
 
-#define BITLOCKER_ENABLE_DEBUG 1
+#define BITLOCKER_ENABLE_DEBUG 0
+#define KERNEL_NAME "opencl_bitlocker_attack"
+//Start size for gws
+#define SEED 1024
 
-static cl_kernel prepare_kernel;
+static cl_kernel block_kernel;
 static struct fmt_main *self;
 
 static cl_mem salt_d, padding_d, w_blocks_d, deviceEncryptedVMK,
-       devicePassword, devicePasswordSize, deviceFound;
+       devicePassword, devicePasswordSize, deviceFound, numPasswordsKernelDev;
 static cl_int ciErr1;
-static size_t szGlobalWorkSize, szLocalWorkSize, maxDeviceWorkgroupSize;
+static cl_int cl_error;
+
 static unsigned int *w_blocks_h;
 static unsigned char salt[BITLOCKER_SALT_SIZE], nonce[BITLOCKER_NONCE_SIZE],
        encryptedVMK[BITLOCKER_VMK_SIZE];
-static unsigned char *tmpIV, *inbuffer, *outbuffer;
+static unsigned char *tmpIV, *inbuffer;
 static int *inbuffer_size;
 
 //static FILE *diskImage;
-static int *hostFound, passwordBufferSize, numPassword, totPsw, i;
-static int max_compute_units = 0;
-static int num_pass_per_thread = 0;
+static int *hostFound, totPsw, i;
 static 	unsigned int tmp_global, IV0, IV4, IV8, IV12;
 
-static int var_max_keys_per_crypt = MAX_KEYS_PER_CRYPT;
+static int * numPasswordsKernel;
+static int salt_done=0;
 
-static struct fmt_tests BitLocker_tests[] = {
+static struct fmt_tests tests[] = {
 	{"$bitlocker$b0599ad6c6a1cf0103000000$0a8b9d0655d3900e9f67280adc27b5d7$033a16cb", "paperino"},
 	{NULL}
+};
+
+static const char * warn[] = {
+        "pass xfer: "  ,  ", crypt: "    ,  ", result xfer: "
 };
 
 static void w_block_precomputed(unsigned char *salt);
 
 #include "opencl-autotune.h"
+#include "memdbg.h"
+
+/* ------- Helper functions ------- */
+static size_t get_task_max_work_group_size()
+{
+	return autotune_get_task_max_work_group_size(FALSE, 0, crypt_kernel);
+}
+
+static void create_clobj(size_t gws, struct fmt_main *self)
+{
+	size_t in_size = BITLOCKER_MAX_INPUT_PASSWORD_LEN * gws;
+
+#define CLCREATEBUFFER(_flags, _size, _string)\
+	clCreateBuffer(context[gpu_id], _flags, _size, NULL, &cl_error);\
+	HANDLE_CLERROR(cl_error, _string);
+
+#define CLKERNELARG(kernel, id, arg, msg)\
+	HANDLE_CLERROR(clSetKernelArg(kernel, id, sizeof(arg), &arg), msg);
+
+
+	inbuffer = (unsigned char *)mem_calloc(in_size, sizeof(unsigned char));
+	inbuffer_size = (int *)mem_calloc(MAX_KEYS_PER_CRYPT, sizeof(int));
+	hostFound = (int *)mem_calloc(1, sizeof(int));
+	numPasswordsKernel = (int *)mem_calloc(1, sizeof(int));
+
+	w_blocks_h = (unsigned int *)mem_calloc((BITLOCKER_SINGLE_BLOCK_SHA_SIZE *
+	                                     BITLOCKER_ITERATION_NUMBER), sizeof(unsigned int));
+
+	deviceEncryptedVMK = CLCREATEBUFFER(CL_MEM_READ_WRITE,
+	                   					VMK_DECRYPT_SIZE * BITLOCKER_MAX_INPUT_PASSWORD_LEN * sizeof(unsigned char),
+	                   					"Cannot allocate vmk");
+
+	devicePassword = CLCREATEBUFFER(CL_MEM_READ_ONLY,
+	                   					in_size * sizeof(unsigned char),
+	                   					"Cannot allocate inbuffer");
+
+	devicePasswordSize = CLCREATEBUFFER(CL_MEM_READ_ONLY,
+	                   					in_size * sizeof(unsigned char),
+	                   					"Cannot allocate inbuffer size");
+
+	deviceFound = CLCREATEBUFFER(CL_MEM_WRITE_ONLY,
+	                   					sizeof(int),
+	                   					"Cannot allocate device found");
+
+	w_blocks_d = CLCREATEBUFFER(CL_MEM_READ_WRITE,
+	                   					BITLOCKER_SINGLE_BLOCK_SHA_SIZE * BITLOCKER_ITERATION_NUMBER * sizeof(unsigned int),
+	                   					"Cannot allocate w blocks");
+
+	numPasswordsKernelDev = CLCREATEBUFFER(CL_MEM_READ_ONLY,
+	                   					sizeof(int),
+	                   					"Cannot allocate numPasswordsKernelDev");
+
+
+	salt_d = CLCREATEBUFFER(CL_MEM_READ_ONLY,
+           					BITLOCKER_SALT_SIZE * sizeof(char),
+           					"Cannot allocate salt_d");
+
+	printf("salt_d %p allocato %zu byte\n", salt_d, BITLOCKER_SALT_SIZE * sizeof(char));
+	padding_d = CLCREATEBUFFER(CL_MEM_READ_ONLY,
+           					BITLOCKER_PADDING_SIZE * sizeof(char),
+           					"Cannot allocate padding_d");
+
+	CLKERNELARG(crypt_kernel, 0, numPasswordsKernelDev, "Error while setting numPasswordsKernelDev");
+	CLKERNELARG(crypt_kernel, 1, devicePassword, "Error while setting numPasswordsKernelDev");
+	CLKERNELARG(crypt_kernel, 2, devicePasswordSize, "Error while setting numPasswordsKernelDev");
+	CLKERNELARG(crypt_kernel, 3, deviceFound, "Error while setting numPasswordsKernelDev");
+	CLKERNELARG(crypt_kernel, 4, deviceEncryptedVMK, "Error while setting numPasswordsKernelDev");
+	CLKERNELARG(crypt_kernel, 5, w_blocks_d, "Error while setting numPasswordsKernelDev");
+	CLKERNELARG(crypt_kernel, 6, IV0, "Error while setting numPasswordsKernelDev");
+	CLKERNELARG(crypt_kernel, 7, IV4, "Error while setting numPasswordsKernelDev");
+	CLKERNELARG(crypt_kernel, 8, IV8, "Error while setting numPasswordsKernelDev");
+	CLKERNELARG(crypt_kernel, 9, IV12, "Error while setting numPasswordsKernelDev");
+
+	memset(inbuffer, '\0', in_size);
+/*
+	CLKERNELARG(block_kernel, 0, salt_d, "Error while setting salt_d");
+	CLKERNELARG(block_kernel, 1, padding_d, "Error while setting padding_d");
+	CLKERNELARG(block_kernel, 2, w_blocks_d, "Error while setting w_blocks_d");
+*/
+}
+
+static void release_clobj(void)
+{
+	if (deviceFound) {
+		printf("release_clobj\n");
+		//HANDLE_CLERROR(clEnqueueUnmapMemObject(queue[gpu_id], pinned_in, inbuffer, 0, NULL, NULL), "Error Unmapping inbuffer");
+		//HANDLE_CLERROR(clEnqueueUnmapMemObject(queue[gpu_id], pinned_out, outbuffer, 0, NULL, NULL), "Error Unmapping outbuffer");
+		//HANDLE_CLERROR(clFinish(queue[gpu_id]), "Error releasing memory mappings");
+
+		HANDLE_CLERROR(clReleaseMemObject(deviceEncryptedVMK), "Release pinned_in");
+		HANDLE_CLERROR(clReleaseMemObject(devicePassword), "Release mem_in");
+		HANDLE_CLERROR(clReleaseMemObject(devicePasswordSize), "Release mem_salt");
+		HANDLE_CLERROR(clReleaseMemObject(deviceFound), "Release pinned_out");
+		HANDLE_CLERROR(clReleaseMemObject(numPasswordsKernelDev), "Release mem_out");
+
+		//free(inbuffer);
+		//free(inbuffer_size);
+		//free(hostFound);
+		//free(numPasswordsKernel);
+
+		deviceFound = NULL;
+	}
+}
+
+static void done(void)
+{
+	if (autotuned) {
+		release_clobj();
+		HANDLE_CLERROR(clReleaseMemObject(w_blocks_d), "Release mem_out");
+		HANDLE_CLERROR(clReleaseKernel(crypt_kernel), "Release kernel");
+		HANDLE_CLERROR(clReleaseProgram(program[gpu_id]), "Release Program");
+
+		autotuned--;
+	}
+}
 
 static void reset(struct db_main *db)
 {
-	char opt[1024];
+	if (!autotuned) {
+		char build_opts[512];
+
+		printf("reset\n");
+		opencl_init("$JOHN/kernels/bitlocker_kernel.cl", gpu_id, build_opts);
+
+		//opencl_build_kernel("$JOHN/kernels/bitlocker_kernel.cl", gpu_id, build_opts, 0);
+
+		///Create Kernel
+		crypt_kernel = clCreateKernel(program[gpu_id], "opencl_bitlocker_attack", &ret_code);
+		HANDLE_CLERROR(ret_code, "Error while creating kernel");
+
+		block_kernel = clCreateKernel(program[gpu_id], "opencl_bitlocker_wblocks", &ret_code);
+		HANDLE_CLERROR(ret_code, "Error creating kernel_prepare. Double-check kernel name?");
+
+		//Initialize openCL tuning (library) for this format.
+		opencl_init_auto_setup(SEED, 0, NULL, warn, 1, self,
+		                       create_clobj, release_clobj,
+		                       BITLOCKER_SINGLE_BLOCK_SHA_SIZE * BITLOCKER_ITERATION_NUMBER * sizeof(int), 0, db);
+
+		//Auto tune execution from shared/included code.
+		autotune_run(self, 1000, 0, 500);
+	}
+#if 0
+	char fileNameAttack[] = "$JOHN/kernels/bitlocker_kernel.cl", opt[1024];
 	size_t deviceGlobalMem = 0;
 	long int globalMemRequired = 0;
 
 	num_pass_per_thread = MAX_PASSWORD_THREAD;
 
-	if (gpu_nvidia(device_info[gpu_id])) {
-		if (nvidia_sm_5x(device_info[gpu_id]))
-			snprintf(opt, sizeof(opt), "-D DEV_NVIDIA_SM50=1"); //-cl-nv-verbose
-		else
-			snprintf(opt, sizeof(opt), "-D DEV_NVIDIA_SM50=0");
-	} else
-		snprintf(opt, sizeof(opt), "-D DEV_NVIDIA_SM50=0");
 
-	opencl_build_kernel("$JOHN/kernels/bitlocker_kernel.cl", gpu_id, opt, 0);
+
+	opencl_build_kernel(fileNameAttack, gpu_id, opt, 0);
 	crypt_kernel =
 	    clCreateKernel(program[gpu_id], "opencl_bitlocker_attack", &ciErr1);
 	HANDLE_CLERROR(ciErr1, "clCreateKernel");
@@ -167,9 +311,10 @@ static void reset(struct db_main *db)
 		("Error global memory size! Required: %ld, Available: %ld GPU_ID: %d\n",
 		 globalMemRequired, deviceGlobalMem, gpu_id);
 
-	db->format->params.max_keys_per_crypt = szGlobalWorkSize;
+	db->format->params.max_keys_per_crypt =
+	    szGlobalWorkSize;   //*num_pass_per_thread;
 	var_max_keys_per_crypt = szGlobalWorkSize;
-	numPassword = szGlobalWorkSize;
+	numPassword = szGlobalWorkSize; //num_pass_per_thread;
 
 	deviceEncryptedVMK =
 	    clCreateBuffer(context[gpu_id], CL_MEM_READ_WRITE,
@@ -200,6 +345,7 @@ static void reset(struct db_main *db)
 	                   BITLOCKER_SINGLE_BLOCK_SHA_SIZE * BITLOCKER_ITERATION_NUMBER *
 	                   sizeof(unsigned int), NULL, &ciErr1);
 	HANDLE_CLERROR(ciErr1, "clCreateBuffer");
+#endif
 	return;
 }
 
@@ -207,38 +353,19 @@ static void reset(struct db_main *db)
 static void init(struct fmt_main *_self)
 {
 	self = _self;
-
 	opencl_prepare_dev(gpu_id);
-	max_compute_units = get_max_compute_units(gpu_id);
-	maxDeviceWorkgroupSize = get_device_max_lws(gpu_id);
-
-	numPassword =
-	    (int)maxDeviceWorkgroupSize * max_compute_units *
-	    BITLOCKER_MAX_INPUT_PASSWORD_LEN;
-	passwordBufferSize =
-	    (int)numPassword * BITLOCKER_MAX_INPUT_PASSWORD_LEN *
-	    sizeof(unsigned char);
-
-	inbuffer =
-	    (unsigned char *)calloc(passwordBufferSize, sizeof(unsigned char));
-	inbuffer_size = (int *)calloc(MAX_KEYS_PER_CRYPT, sizeof(int));
-	hostFound = (int *)calloc(1, sizeof(int));
-	outbuffer =
-	    (unsigned char *)calloc(BITLOCKER_MAX_INPUT_PASSWORD_LEN + 2,
-	                            sizeof(unsigned char));
 }
 
 void w_block_precomputed(unsigned char *salt)
 {
-	char *opt = NULL;
 	unsigned char *padding;
 	uint64_t msgLen;
-	cl_mem w_blocks_d_local;
 
 	if (salt == NULL)
 		return;
 
-	padding = (unsigned char *)calloc(BITLOCKER_PADDING_SIZE, sizeof(unsigned char));
+	padding =
+	    (unsigned char *)calloc(BITLOCKER_PADDING_SIZE, sizeof(unsigned char));
 	padding[0] = 0x80;
 	memset(padding + 1, 0, 31);
 	msgLen = (BITLOCKER_FIXED_PART_INPUT_CHAIN_HASH << 3);
@@ -246,6 +373,7 @@ void w_block_precomputed(unsigned char *salt)
 		padding[BITLOCKER_PADDING_SIZE - 1 - i] =
 		    (uint8_t)(msgLen >> (i * 8));
 
+/*
 	salt_d =
 	    clCreateBuffer(context[gpu_id], CL_MEM_READ_ONLY,
 	                   BITLOCKER_SALT_SIZE * sizeof(unsigned char), NULL, &ciErr1);
@@ -260,48 +388,62 @@ void w_block_precomputed(unsigned char *salt)
 	                            BITLOCKER_SINGLE_BLOCK_SHA_SIZE * BITLOCKER_ITERATION_NUMBER *
 	                            sizeof(int), NULL, &ciErr1);
 	HANDLE_CLERROR(ciErr1, "clCreateBuffer");
+*/
 
-	w_blocks_h = (unsigned int *)calloc((BITLOCKER_SINGLE_BLOCK_SHA_SIZE *
-	                                     BITLOCKER_ITERATION_NUMBER), sizeof(unsigned int));
-	if (!w_blocks_h)
-		goto out;
+//	opencl_build_kernel("$JOHN/kernels/wblock_kernel.cl", gpu_id, NULL, 0);
 
-	opencl_build_kernel("$JOHN/kernels/bitlocker_kernel.cl", gpu_id, opt, 0);
-	prepare_kernel =
-	    clCreateKernel(program[gpu_id], "opencl_bitlocker_wblocks", &ciErr1);
-	HANDLE_CLERROR(ciErr1,
-	               "Error creating kernel_prepare. Double-check kernel name?");
+		printf("w_block_precomputed 0, salt[0]_ %x\n", salt[0]);
+
+
+	printf("salt_d %p write %zu byte\n", salt_d, BITLOCKER_SALT_SIZE * sizeof(char));
+
 
 	ciErr1 =
 	    clEnqueueWriteBuffer(queue[gpu_id], salt_d, CL_TRUE, 0,
-	                         BITLOCKER_SALT_SIZE * sizeof(unsigned char), salt, 0, NULL, NULL);
+	                         BITLOCKER_SALT_SIZE * sizeof(char), salt, 0, NULL, NULL);
 	HANDLE_CLERROR(ciErr1, "clEnqueueWriteBuffer");
+
+	printf("w_block_precomputed 1\n");
+
+
 	ciErr1 =
 	    clEnqueueWriteBuffer(queue[gpu_id], padding_d, CL_TRUE, 0,
-	                         BITLOCKER_PADDING_SIZE * sizeof(unsigned char), padding, 0, NULL,
-	                         NULL);
+	                         BITLOCKER_PADDING_SIZE * sizeof(char), padding, 0, NULL, NULL);
 	HANDLE_CLERROR(ciErr1, "clEnqueueWriteBuffer");
 
+	printf("w_block_precomputed 3\n");
 
+/*
 	ciErr1 =
-	    clSetKernelArg(prepare_kernel, 0, sizeof(cl_mem), (void *)&salt_d);
+	    clSetKernelArg(block_kernel, 0, sizeof(cl_int), (void *)&iter_num);
+	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
+*/
+	/*
+	ciErr1 =
+	    clSetKernelArg(block_kernel, 0, sizeof(cl_mem), (void *)&salt_d);
 	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
 	ciErr1 =
-	    clSetKernelArg(prepare_kernel, 1, sizeof(cl_mem), (void *)&padding_d);
+	    clSetKernelArg(block_kernel, 1, sizeof(cl_mem), (void *)&padding_d);
 	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
 	ciErr1 =
-	    clSetKernelArg(prepare_kernel, 2, sizeof(cl_mem), (void *)&w_blocks_d_local);
+	    clSetKernelArg(block_kernel, 2, sizeof(cl_mem), (void *)&w_blocks_d);
 	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
 
-	szLocalWorkSize = get_kernel_max_lws(gpu_id, prepare_kernel);
+	max_compute_units = get_max_compute_units(gpu_id);
+	szLocalWorkSize = get_kernel_max_lws(gpu_id, block_kernel);
 	szGlobalWorkSize = szLocalWorkSize * max_compute_units;
 
+printf("dopo di szLocalWorkSize: %zu, szGlobalWorkSize: %zu\n",
+	szLocalWorkSize, szGlobalWorkSize);
+*/
+
 	ciErr1 =
-	    clEnqueueNDRangeKernel(queue[gpu_id], prepare_kernel, 1, NULL,
-	                           &szGlobalWorkSize, &szLocalWorkSize, 0, NULL, NULL);
+	    clEnqueueNDRangeKernel(queue[gpu_id], block_kernel, 1, NULL,
+	                           &global_work_size, &local_work_size,
+	                           0, NULL, NULL);
 	HANDLE_CLERROR(ciErr1, "clEnqueueNDRangeKernel");
 
-	ciErr1 = clEnqueueReadBuffer(queue[gpu_id], w_blocks_d_local, CL_TRUE, 0,
+	ciErr1 = clEnqueueReadBuffer(queue[gpu_id], w_blocks_d, CL_TRUE, 0,
 	                             BITLOCKER_SINGLE_BLOCK_SHA_SIZE * BITLOCKER_ITERATION_NUMBER *
 	                             sizeof(unsigned int), w_blocks_h, 0, NULL, NULL);
 	HANDLE_CLERROR(ciErr1, "clEnqueueNDRangeKernel");
@@ -309,13 +451,10 @@ void w_block_precomputed(unsigned char *salt)
 	HANDLE_CLERROR(clFlush(queue[gpu_id]), "clFlush");
 	HANDLE_CLERROR(clFinish(queue[gpu_id]), "clFinish");
 
-out:
+	//HANDLE_CLERROR(clReleaseKernel(block_kernel), "clReleaseKernel");
 
-	HANDLE_CLERROR(clReleaseKernel(prepare_kernel), "clReleaseKernel");
-
-	clReleaseMemObject(salt_d);
+	//clReleaseMemObject(salt_d);
 	clReleaseMemObject(padding_d);
-	clReleaseMemObject(w_blocks_d_local);
 	free(padding);
 
 	return;
@@ -352,16 +491,18 @@ static int valid(char *ciphertext, struct fmt_main *self)
 	return 1;
 }
 
-static void *get_binary(char *ciphertext)
-{
-	return ciphertext;
-}
-
 static void *get_salt(char *ciphertext)
 {
 	int i = 0;
 	char *hash_format;
 	char *p;
+
+	memset(salt, 0, sizeof(salt));
+	printf("get_salt, salt_done: %d\n", salt_done);
+	if (salt_done > 0)
+		return salt;
+
+	salt_done=1;
 
 	hash_format = strdup(ciphertext);
 	hash_format += FORMAT_TAG_LEN;
@@ -369,7 +510,7 @@ static void *get_salt(char *ciphertext)
 	p = strtokm(hash_format, "$");
 	if (strlen(p) != BITLOCKER_NONCE_SIZE * 2)
 		return NULL;
-		          
+
 	for (i = 0; i < BITLOCKER_NONCE_SIZE; i++) {
 		nonce[i] =
 		    (p[2 * i] <=
@@ -414,11 +555,26 @@ static void *get_salt(char *ciphertext)
 #endif
 	}
 
-	w_block_precomputed(salt);
+	return salt;
+}
+
+
+static void set_salt(void *psalt)
+{
+	unsigned char * local_salt = (unsigned char *) salt;
+
+	if (salt_done == 2)
+		return;
+
+	salt_done=2;
+
+	printf("salt_done: %d set local_salt: %x - %x\n", salt_done, local_salt[0], local_salt[1]);
+	w_block_precomputed(local_salt);
 	if (!w_blocks_h) {
 		error_msg("Error... Exit\n");
 	}
 
+	printf("w_blocks_h: %x\n", w_blocks_h[0]);
 	tmpIV = (unsigned char *)calloc(BITLOCKER_IV_SIZE, sizeof(unsigned char));
 
 	memset(tmpIV, 0, BITLOCKER_IV_SIZE);
@@ -429,7 +585,7 @@ static void *get_salt(char *ciphertext)
 	*tmpIV = (unsigned char)(BITLOCKER_IV_SIZE - 1 - BITLOCKER_NONCE_SIZE - 1);
 	tmpIV[BITLOCKER_IV_SIZE - 1] = 1;
 
-	tmp_global = 
+	tmp_global =
 	(((unsigned int *)(tmpIV)) [0]);
 	IV0 =
 	    (unsigned int)(((unsigned int)(tmp_global & 0xff000000)) >> 24) |
@@ -457,9 +613,6 @@ static void *get_salt(char *ciphertext)
 	    (unsigned int)((unsigned int)(tmp_global & 0x00ff0000) >> 8) |
 	    (unsigned int)((unsigned int)(tmp_global & 0x0000ff00) << 8) |
 	    (unsigned int)((unsigned int)(tmp_global & 0x000000ff) << 24);
-
-
-	return tmpIV;
 }
 
 static void set_key(char *key, int index)
@@ -485,95 +638,77 @@ static char *get_key(int index)
 	return ret;
 }
 
-static int crypt_all(int *pcount, struct db_salt *salt)
+static int crypt_all(int *count, struct db_salt *salt)
 {
-	int numPassword = *pcount;
+	int numPassword[1];
+	size_t *lws = local_work_size ? &local_work_size : NULL;
+
+	numPassword[0] = *count;
+	global_work_size = GET_MULTIPLE_OR_BIGGER(numPassword[0], local_work_size);
 
 #if BITLOCKER_ENABLE_DEBUG == 1
 	time_t start, end;
 	double dif;
+	printf("\n[BitCracker] -> Starting Attack, #Passwords: %d\n", numPassword[0]);
+	//, , Global Work Size: %zu, Local Work Size: %zu szGlobalWorkSize, szLocalWorkSize);
 #endif
 
+printf("crypt all, numPassword: %d\n", numPassword[0]);
+	ciErr1 =
+	    clEnqueueWriteBuffer(queue[gpu_id], numPasswordsKernelDev, CL_TRUE, 0,
+	                         sizeof(int), numPassword, 0, NULL, multi_profilingEvent[0]);
+	HANDLE_CLERROR(ciErr1, "clEnqueueWriteBuffer");
 
-#if BITLOCKER_ENABLE_DEBUG == 1
-	printf
-	("\n[BitCracker] -> Starting Attack, #Passwords: %d, Global Work Size: %zu, Local Work Size: %zu\n",
-	 numPassword, szGlobalWorkSize, szLocalWorkSize);
-#endif
 
+/*
 	ciErr1 =
 	    clEnqueueWriteBuffer(queue[gpu_id], w_blocks_d, CL_TRUE, 0,
 	                         BITLOCKER_SINGLE_BLOCK_SHA_SIZE * BITLOCKER_ITERATION_NUMBER *
-	                         sizeof(int), w_blocks_h, 0, NULL, NULL);
+	                         sizeof(int), w_blocks_h, 0, NULL, multi_profilingEvent[1]);
 	HANDLE_CLERROR(ciErr1, "clEnqueueWriteBuffer");
+*/
 
 	ciErr1 =
 	    clEnqueueWriteBuffer(queue[gpu_id], deviceEncryptedVMK, CL_TRUE, 0,
-	                         BITLOCKER_VMK_SIZE * sizeof(char), encryptedVMK, 0, NULL, NULL);
+	                         BITLOCKER_VMK_SIZE * sizeof(char), encryptedVMK, 0, NULL, multi_profilingEvent[2]);
 	HANDLE_CLERROR(ciErr1, "clEnqueueWriteBuffer");
 
 
 	ciErr1 =
 	    clEnqueueWriteBuffer(queue[gpu_id], devicePassword, CL_TRUE, 0,
-	                         numPassword * BITLOCKER_MAX_INPUT_PASSWORD_LEN, inbuffer, 0, NULL, NULL);
+	                         numPassword[0] * BITLOCKER_MAX_INPUT_PASSWORD_LEN, inbuffer, 0, NULL, multi_profilingEvent[3]);
 	HANDLE_CLERROR(ciErr1, "clEnqueueWriteBuffer");
 
 	ciErr1 =
 	    clEnqueueWriteBuffer(queue[gpu_id], devicePasswordSize, CL_TRUE, 0,
-	                         numPassword * sizeof(int), inbuffer_size, 0, NULL, NULL);
+	                         numPassword[0] * sizeof(int), inbuffer_size, 0, NULL, multi_profilingEvent[4]);
 	HANDLE_CLERROR(ciErr1, "clEnqueueWriteBuffer");
 
 	hostFound[0] = -1;
 	ciErr1 =
 	    clEnqueueWriteBuffer(queue[gpu_id], deviceFound, CL_TRUE, 0,
-	                         sizeof(int), hostFound, 0, NULL, NULL);
+	                         sizeof(int), hostFound, 0, NULL, multi_profilingEvent[5]);
 	HANDLE_CLERROR(ciErr1, "clEnqueueWriteBuffer");
-
-	ciErr1 =
-	    clSetKernelArg(crypt_kernel, 0, sizeof(cl_int), (void *)&numPassword);
-	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
-
-	ciErr1 |=
-	    clSetKernelArg(crypt_kernel, 1, sizeof(cl_mem),
-	                   (void *)&devicePassword);
-	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
-
-	ciErr1 |=
-	    clSetKernelArg(crypt_kernel, 2, sizeof(cl_mem),
-	                   (void *)&devicePasswordSize);
-	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
-
-	ciErr1 |=
-	    clSetKernelArg(crypt_kernel, 3, sizeof(cl_mem), (void *)&deviceFound);
-	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
-
-	ciErr1 |=
-	    clSetKernelArg(crypt_kernel, 4, sizeof(cl_mem),
-	                   (void *)&deviceEncryptedVMK);
-	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
-
-	ciErr1 |=
-	    clSetKernelArg(crypt_kernel, 5, sizeof(cl_mem), (void *)&w_blocks_d);
-	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
-
-	ciErr1 |= clSetKernelArg(crypt_kernel, 6, sizeof(cl_int), (void *)&IV0);
-	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
-
-	ciErr1 |= clSetKernelArg(crypt_kernel, 7, sizeof(cl_int), (void *)&IV4);
-	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
-
-	ciErr1 |= clSetKernelArg(crypt_kernel, 8, sizeof(cl_int), (void *)&IV8);
-	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
-
-	ciErr1 |= clSetKernelArg(crypt_kernel, 9, sizeof(cl_int), (void *)&IV12);
-	HANDLE_CLERROR(ciErr1, "clSetKernelArg");
 
 #if BITLOCKER_ENABLE_DEBUG == 1
 	time(&start);
 #endif
 
+	//Run kernel
+	BENCH_CLERROR(clEnqueueNDRangeKernel(queue[gpu_id], crypt_kernel, 1,
+		NULL, &global_work_size, lws, 0, NULL, multi_profilingEvent[6]),
+		"Set ND range");
+	BENCH_CLERROR(clEnqueueReadBuffer(queue[gpu_id], deviceFound, CL_TRUE, 0,
+					sizeof(int), hostFound, 0, NULL, multi_profilingEvent[7]), "Copy data back");
 
-	ciErr1 =
+	//Await completion of all the above
+	BENCH_CLERROR(clFinish(queue[gpu_id]), "clFinish error");
+
+
+
+#if 0
+
+		ciErr1 =
 	    clEnqueueNDRangeKernel(queue[gpu_id], crypt_kernel, 1, NULL,
 	                           &szGlobalWorkSize, &szLocalWorkSize, 0, NULL, NULL);
 	HANDLE_CLERROR(ciErr1, "clEnqueueNDRangeKernel");
@@ -583,6 +718,7 @@ static int crypt_all(int *pcount, struct db_salt *salt)
 	                        sizeof(unsigned int), hostFound, 0, NULL, NULL);
 	HANDLE_CLERROR(ciErr1, "clEnqueueNDRangeKernel");
 	HANDLE_CLERROR(clFlush(queue[gpu_id]), "clFlush");
+#endif
 
 #if BITLOCKER_ENABLE_DEBUG == 1
 
@@ -591,15 +727,15 @@ static int crypt_all(int *pcount, struct db_salt *salt)
 
 	printf
 	("[BitCracker] -> Attack stats: %d passwords evaluated in %.2lf seconds => %.2f pwd/s\n",
-	 numPassword, dif, (double)(numPassword / dif));
+	 numPassword[0], dif, (double)(numPassword[0] / dif));
 
 #endif
 
-	HANDLE_CLERROR(clFinish(queue[gpu_id]), "clFinish");
+	//HANDLE_CLERROR(clFinish(queue[gpu_id]), "clFinish");
 
-	totPsw += numPassword;
+	totPsw += numPassword[0];
 
-	return numPassword;
+	return numPassword[0];
 }
 
 
@@ -629,31 +765,7 @@ static int cmp_one(void *binary, int index)
 
 static int cmp_exact(char *source, int index)
 {
-	if (hostFound[0] == index)
-		return 1;
-	else
-		return 0;
-}
-
-static void done(void)
-{
-	if (w_blocks_d)
-		HANDLE_CLERROR(clReleaseMemObject(w_blocks_d), "clReleaseMemObject");
-	if (devicePassword)
-		HANDLE_CLERROR(clReleaseMemObject(devicePassword),
-		               "clReleaseMemObject");
-	if (devicePasswordSize)
-		HANDLE_CLERROR(clReleaseMemObject(devicePasswordSize),
-		               "clReleaseMemObject");
-	if (deviceEncryptedVMK)
-		HANDLE_CLERROR(clReleaseMemObject(deviceEncryptedVMK),
-		               "clReleaseMemObject");
-	if (deviceFound)
-		HANDLE_CLERROR(clReleaseMemObject(deviceFound), "clReleaseMemObject");
-	if (crypt_kernel)
-		HANDLE_CLERROR(clReleaseKernel(crypt_kernel), "clReleaseKernel");
-
-	return;
+	return 1;
 }
 
 struct fmt_main fmt_opencl_bitlocker = {
@@ -673,13 +785,11 @@ struct fmt_main fmt_opencl_bitlocker = {
 		MIN_KEYS_PER_CRYPT,
 		MAX_KEYS_PER_CRYPT,
 		FMT_CASE | FMT_8_BIT,
-		{NULL}
-/*		,
+		{NULL},
 		{
 			FORMAT_TAG
 		},
-		BitLocker_tests
-*/
+		tests
 	}, {
 		init,
 		done,
@@ -687,7 +797,7 @@ struct fmt_main fmt_opencl_bitlocker = {
 		fmt_default_prepare,
 		valid,
 		fmt_default_split,  //split
-		get_binary,
+		fmt_default_binary,
 		get_salt,
 		{NULL},
 		fmt_default_source,
@@ -702,7 +812,7 @@ struct fmt_main fmt_opencl_bitlocker = {
 		},
 		fmt_default_salt_hash,  //salt_hash
 		NULL,
-		fmt_default_set_salt,
+		set_salt,
 		set_key,
 		get_key,
 		fmt_default_clear_keys,
